1 solve test may fail problem 2 refactor etcd package dir and name
